requirements.txt
python-telegram-bot
google-generativeai
pymongo
flask